services/product-finder/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.117.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
lxml==6.0.2
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.20.1
selenium>=4.25.0
webdriver-manager>=4.0.2
